requirements.txt
amqp==5.1.1
asgiref==3.7.2
astroid==2.15.5
billiard==4.1.0
celery==5.3.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
dill==0.3.6
Django==4.2.2
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-filter==23.2
django-timezone-field==5.1
djangorestframework==3.14.0
drf-yasg==1.21.6
inflection==0.5.1
isort==5.12.0
kombu==5.3.1
lazy-object-proxy==1.9.0
Markdown==3.4.3
mccabe==0.7.0
packaging==23.1
Pillow==9.5.0
platformdirs==3.7.0
prompt-toolkit==3.0.38
pylint==2.17.4
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.4
tomlkit==0.11.8
tzdata==2023.3
uritemplate==4.1.1
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
